backend/requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-text-splitters
transformers
torch
sentence-transformers
huggingface-hub
google-api-python-client
google-ai-generativelanguage
google-generativeai
pandas
numpy
scikit-learn
scipy
nltk
requests
python-dotenv
faiss-cpu
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
pydantic
python-jose
pypdfium2
PyJWT
itsdangerous
Jinja2
python-multipart
google-cloud-secret-manager
slowapi
click==8.1.7
